+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -16,6 +16,8 @@ group: Deployment & Configs
 
 - Added 'Pick Elements' component.
 - Added 'Pick Points' component.
+- Renamed 'Document Warnings' -> 'Query Warnings'.
+- Added context menu to 'Built-In Failure Definitions' to filter by severity.
 
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}